Dala 395 dummy qa and mq integration wrapper (#189)
wrap exception handling for rabbit listeners
